refactor: git logic to separate file
